- dynamic keyboard for payments

+++ lib/feature/payment/request_from_new_recipient/request_from_new_recipient_view_model.dart
@@ -208,9 +208,10 @@ class RequestFromNewRecipientViewModel extends BasePageViewModel {
             ibanOrMobile: ibanOrMobileController.text,
             purpose: purposeController.text,
             purposeDetail: purposeDetailController.text,
-            amount: int.parse(ProviderScope.containerOf(context)
-                .read(requestMoneyViewModelProvider)
-                .currentPinValue),
+            amount: double.parse(ProviderScope.containerOf(context)
+                    .read(requestMoneyViewModelProvider)
+                    .currentPinValue)
+                .toInt(),
             limit: limit,
             dbtrBic: dbtrBic ?? "",
             dbtrAcct: dbtrAcct ?? "",
